requirements.txt.txt
python-telegram-bot==20.0
Flask==2.3.3